Final response to ch 3 quiz question 1
The variable x was not retaining user input. Issue was that it returned the	result but was not storing it in anything like a variable.
	- I changed the program to initialize the variable 'x' by calling the 'readNumber' function immediately.
The next issue was that the function was taking an integer as a parameter unnecessarily.
	- I removed the parameter.
The last issue is that the local variable in the 'readNumber' function was not initialized.
	- I initialized the variable.