Change program to use implicit type conversion instead of explicit
Instead of using static_case<TYPE>(VAR) I use a function to take int parameters
but use it with a char parameter. The function then returns the parameter making
no change. The fact the function only takes int params causes it to implicitly
type cast the char input as an int.