Experiment with std::string_view initialization and assignment
The example was trying to illustrate that the original variable is not
modified when string_view is assigned a different value. This brought
up the question of whether or not this behaves the same way even if you
initialize the std::string_view with a literal instead of via a variable.

It was unclear to me if the initialized value was different than the assigned
value in that it had to exist independent of the string_view but I can see
now that it does not.