Second attempt to use string and string_view literal suffixes directly
Attemped to use the scope resolution operator to use the suffixes without
importing the whole namespace but this time as the name of the header file.
This also gave the same error as before:
> ```
> string_view_literals.cpp: In function ‘int main()’:
> string_view_literals.cpp:11:18: error: unable to find string literal operator ‘operator""std’ with ‘const char [5]’, ‘long unsigned int’ arguments
>    11 |     std::cout << "goo\n"std::string::s;  // s suffix is a std::string literal
>       |                  ^~~~~~~~~~
> string_view_literals.cpp:12:18: error: unable to find string literal operator ‘operator""std’ with ‘const char [5]’, ‘long unsigned int’ arguments
>    12 |     std::cout << "moo\n"std::string_view::sv; // sv suffix is a std::string_view literal
>       |                  ^~~~~~~~~~
> ```

I'm not sure why it doesn't work. I guess I am simply missing information about
how namespaces work. I will wait for that lesson and possibly revisit this.