Add additional information in the comments
The following lesson (5.9) covered the idea of ownership and improper
use of std::string_view. It reminded me of this test program I wrote for
practice and I realized I inadvertently avoided improper use. I wanted to
record that fact in the program comments in case I return to it.